v0.7.5: added code for new Tengu abilities

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/effects/Speck.java
@@ -271,7 +271,7 @@ public class Speck extends Image {
 			break;
 			
 		case STEAM:
-			speed.y = -Random.Float( 20, 30 );
+			speed.y = -Random.Float( 10, 15 );
 			angularSpeed = Random.Float( +180 );
 			angle = Random.Float( 360 );
 			lifespan = 1f;
